requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
alembic
passlib
bcrypt==4.0.1
python-multipart
slowapi
pyjwt
email-validator
boto3
psycopg2-binary